fix(ci): satisfy the base-ref craftsmanship gate for the docker_smoke fix
git merge-base HEAD origin/main was 4 commits behind origin/main —
rebased first (per PR #423-425's shared cause), then re-ran the CI's
exact invocation, `check_craftsmanship.py --base origin/main`, which
diffs against the base ref's baseline rather than the working tree's
and caught three real violations `check_craftsmanship.py` alone (no
--base) does not:

- scripts/docker_smoke_client.py::main exceeded the 40-line method cap
  (CLAUDE.md's local tightening of coding-standards.md §4.2) — split
  into _build_parser()/_evaluate()/main(), each under the limit.
- scripts/docker_smoke_client.py::TOOLS_LIST_ID (value 3) had no
  `# source:` comment — added one, same source as PROTOCOL_VERSION
  above it (docker_smoke.sh's original REQUESTS heredoc, where
  tools/list was request id=3).
- tests_py/infrastructure/test_stdio_eof_drain.py grew to 318 lines,
  over the 300-line file cap, after the prior commit appended the
  docker-smoke-specific pinning test to it. Split that test class into
  a new sibling file, test_docker_smoke_stdio_eof_drain.py, which
  imports (not duplicates) _server/_GatedWriteStream/_collect from the
  original — both files now under the cap, no test content lost.

No production logic changed: pytest (1448 passed / 35 skipped),
ruff check/format, and shellcheck all still clean; the deterministic
reproduction (real mcp SDK, docker_smoke's own id=3/tools/list
request, write-then-close loses it / drain-then-close does not) is
unchanged, just relocated.
